feat: Implement user onboarding wizard
Adds an onboarding wizard to guide new users through the application's basic features. The wizard is displayed on the first launch and can be dismissed by the user. It also includes logic to track if onboarding has been completed using local storage. The `Tabs` component is updated to include a new "Recipe" tab icon.